Wrote proper automatic KGO testing

+++ test/testennuf/kgo/test_keras.py
@@ -3,6 +3,7 @@ import shutil
 import subprocess
 
 import numpy as np
+import pytest
 
 import ennuf
 from testennuf import TMPDIR, RANDOM_SEED
@@ -19,7 +20,7 @@ from testennuf.kgo.template import template_test_kgo
 #     keras_model = SimpleMLP.build_functional()
 #     model = from_functional(keras_model)
 #     model.formatter = MinimalistFormatter()
-#     modelpath = test_kgo_dir.joinpath(f'{model.id_}_mod.f90')
+#     modelpath = test_kgo_dir.joinpath(f'{model.name}_mod.f90')
 #     model.create_fortran_module(
 #         modelpath,
 #         overwrite=True,
@@ -42,7 +43,7 @@ from testennuf.kgo.template import template_test_kgo
 #     txt_reader_mod_path = TMPDIR.joinpath('matrix_txt_reader_mod.f90')
 #     neural_net_mod_path = TMPDIR.joinpath('neural_net_mod.f90')
 #     main_path = TMPDIR.joinpath('main.f90')
-#     executablepath = test_kgo_dir.joinpath(f"run_{model.id_}")
+#     executablepath = test_kgo_dir.joinpath(f"run_{model.name}")
 #     f90_files = [modelpath, txt_reader_mod_path, neural_net_mod_path, main_path]
 #     object_files = [f90_file.with_suffix('.o') for f90_file in f90_files]
 #     command_compile_objects = [ennuf.CONFIG.compiler, '-c', *f90_files]
@@ -60,9 +61,10 @@ def template_test_keras_functional(keras_model):
     from ennuf.keras import from_functional
 
     model = from_functional(keras_model)
-    dir_ = TMPDIR.joinpath('keras', f'{model.id_}')
-    shutil.rmtree(dir_)
-    dir_.mkdir()
+    dir_ = TMPDIR.joinpath('keras', f'{model.name}')
+    if dir_.exists():
+        shutil.rmtree(dir_)
+    dir_.mkdir(parents=True)
     template_test_kgo(model, dir_, keras_model.predict)
     shutil.rmtree(dir_)
 
@@ -70,9 +72,15 @@ def template_test_keras_functional(keras_model):
 def test_simple_mlp():
     from testennuf.example_models.simple_mlp import SimpleMLP
 
+    keras_model = SimpleMLP.build_functional_easy()
+    template_test_keras_functional(keras_model)
+
     keras_model = SimpleMLP.build_functional()
     template_test_keras_functional(keras_model)
+
     keras_model = SimpleMLP.build_functional_2()
     template_test_keras_functional(keras_model)
+
     keras_model = SimpleMLP.build_functional_1_1()
-    template_test_keras_functional(keras_model)
+    with pytest.raises(NotImplementedError):
+        template_test_keras_functional(keras_model)
